change type of output value

+++ src/main/java/com/bankingeconomy/hadoop/mapreduce/UserTransactionSumReducer.java
@@ -1,18 +1,19 @@
 package com.bankingeconomy.hadoop.mapreduce;
 
+import org.apache.hadoop.io.DoubleWritable;
 import org.apache.hadoop.io.IntWritable;
 import org.apache.hadoop.io.Text;
 import org.apache.hadoop.mapreduce.Reducer;
 
 import java.io.IOException;
 
-public class UserTransactionSumReducer extends Reducer<Text, IntWritable, Text, IntWritable> {
+public class UserTransactionSumReducer extends Reducer<Text, DoubleWritable, Text, DoubleWritable> {
     @Override
-    public void reduce(Text key, Iterable<IntWritable> values, Context context) throws IOException, InterruptedException {
-        int sum = 0;
+    public void reduce(Text key, Iterable<DoubleWritable> values, Context context) throws IOException, InterruptedException {
+        double sum = 0;
         for(var value : values) {
             sum += value.get();
         }
-        context.write(key, new IntWritable(sum));
+        context.write(key, new DoubleWritable(sum));
     }
 }
